requirements.txt
pandas == 2.2.2
streamlit== 1.37.1
plotly== 5.24.1
google-cloud-bigquery== 3.30.0